refactor: Move cross-validation plots to their own module
