+++ skore/src/skore/sklearn/cross_validation/plots/__init__.py
@@ -1,0 +1,1 @@
+"""Cross-validation plots."""
