+++ skore/src/skore/sklearn/cross_validation/plots/compare_scores_plot.py
