+++ skore/src/skore/sklearn/cross_validation/plots/timing_normalized_plot.py
