+++ skore/src/skore/sklearn/cross_validation/plots/timing_plot.py
